feat : 게이트웨이 테스트를 위한 매핑 경로 수정
